Fix/issue 513

+++ engine/src/gfx/matrix.h
@@ -1,23 +1,28 @@
-/*
- * Vega Strike
- * Copyright (C) 2001-2002 Alan Shieh Rewritten by Daniel Horn for Double precision
+/**
+ * matrix.h
  *
- * http://vegastrike.sourceforge.net/
+ * Copyright (C) 2001-2002 Alan Shieh and Daniel Horn
+ * Copyright (C) 2020-2021 pyramid3d, Stephen G. Tuggy, and other Vega Strike
+ * contributors
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
+ * https://github.com/vegastrike/Vega-Strike-Engine-Source
  *
- * This program is distributed in the hope that it will be useful,
+ * This file is part of Vega Strike.
+ *
+ * Vega Strike is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Vega Strike is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ * along with Vega Strike.  If not, see <https://www.gnu.org/licenses/>.
  */
+
 #ifndef _MATRIX_H
 #define _MATRIX_H
 
@@ -355,8 +360,8 @@ inline void InvertMatrix( Matrix &o, const Matrix &trans )
 
 inline void Rotate( Matrix &tmp, const Vector &axis, float angle )
 {
-    double c = cos( angle );
-    double s = sin( angle );
+    float c = cos( angle );
+    float s = sin( angle );
 //Row, COl
 #define M( a, b ) (tmp.r[b*3+a])
     M( 0, 0 ) = axis.i*axis.i*(1-c)+c;
